feat(flow): integrate per-camera flow tree into health tab, remove flow tab
