+++ web/src/lib/i18n/zh.json
@@ -1731,6 +1731,8 @@
   "flow.buffer": "缓冲",
   "flow.dwell": "驻留 avg/max",
   "flow.noConsumers": "无消费者(仅录制)",
+  "flow.treeTitle": "链路详情",
+  "flow.emptyCamera": "该相机暂无活动流",
   "flow.noViewers": "无观看者",
   "flow.liveLatency": "端到端延迟",
   "health.stability.uptime": "24h 在线率",
